requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.3
google-api-python-client==2.104.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0
requests==2.31.0